+++ db/migrate/20200114185311_create_activity_logs.rb
@@ -5,7 +5,7 @@ class CreateActivityLogs < ActiveRecord::Migration[6.0]
       t.string      :schema_name,   limit: 256,   null: true,   comment: 'Name of schema'
       t.string      :table_name,    limit: 256,   null: true,   comment: 'Name of table'
       t.string      :column_name,   limit: 256,   null: true,   comment: 'Name of column'
-      t.string      :action,        limit: 1024,  null: false,  comment: 'Executed action / activity'
+      t.text        :action,                      null: false,  comment: 'Executed action / activity'
       t.string      :client_ip,     limit: 40,    null: true,   comment: 'Client IP address for request'
       t.timestamps
     end