+++ test/models/user_test.rb
@@ -88,19 +88,18 @@ class UserTest < ActiveSupport::TestCase
   test "increment failed logons" do
     user = User.find(sandro_user.id)
     original_count = user.failed_logons
-    user.increment_failed_logons
+    run_with_current_user { user.increment_failed_logons }
     user = User.find(sandro_user.id)                                            # Reload from DB
     assert_equal original_count+1, user.failed_logons, 'Failed logons should be incremented by 1'
-    user.reset_failed_logons # Restore original state
+    run_with_current_user { user.reset_failed_logons } # Restore original state
   end
 
   test "lock account" do
-    User.find(sandro_user.id).lock_account
+    run_with_current_user { User.find(sandro_user.id).lock_account }
     reloaded_user = User.find(sandro_user.id)
     assert_equal 'Y', reloaded_user.yn_account_locked, 'User should be locked now'
     # Restore original state
-    reloaded_user.yn_account_locked = 'N'
-    reloaded_user.save!
+    run_with_current_user { reloaded_user.update!(yn_account_locked: 'N') }
   end
 
   test "check_for_system_init_completed" do